Bugfix/issue 2234

+++ src/edt/edt/edtPartialService.cc
@@ -762,9 +762,9 @@ PartialShapeFinder::visit_cell (const db::Cell &cell, const db::Box &hit_box, co
 {
   if (! point_mode ()) {
 
-    for (std::vector<int>::const_iterator l = layers ().begin (); l != layers ().end (); ++l) {
+    for (std::vector<unsigned int>::const_iterator l = layers ().begin (); l != layers ().end (); ++l) {
 
-      if (layers ().size () == 1 || (layers ().size () > 1 && cell.bbox ((unsigned int) *l).touches (scan_box))) {
+      if (layers ().size () == 1 || (layers ().size () > 1 && cell.bbox (*l).touches (scan_box))) {
 
         checkpoint ();
 
@@ -897,9 +897,9 @@ PartialShapeFinder::visit_cell (const db::Cell &cell, const db::Box &hit_box, co
 
   } else {
 
-    for (std::vector<int>::const_iterator l = layers ().begin (); l != layers ().end (); ++l) {
+    for (std::vector<unsigned int>::const_iterator l = layers ().begin (); l != layers ().end (); ++l) {
 
-      if (layers ().size () == 1 || (layers ().size () > 1 && cell.bbox ((unsigned int) *l).touches (hit_box))) {
+      if (layers ().size () == 1 || (layers ().size () > 1 && cell.bbox (*l).touches (hit_box))) {
 
         checkpoint ();
